refactor(cst): Rename TreeGen to CstGen
